Break out functions into more easily maintainable file structure. First pass.
